Incremental work on parallel messaging. Some of the receiving node code is now in place. This is the last checkin with the dummy messaging, next round will use MPI. -- Upi
git-svn-id: https://svn.code.sf.net/p/moose/code/moose/trunk@23 d1858a74-f715-0410-99c9-c66fe9cce3d6

+++ scheduling/ClockJobWrapper.cpp
@@ -283,7 +283,7 @@ ClockTickMsgSrc::ClockTickMsgSrc( Element* e, Element* target )
 		conn_( e ), next_( 0 ), target_( target )
 {
  	if ( Ftype1< double >::get( target, "dt", dt_ ) ) {
- 		if ( Ftype1< int >::get( target, "stage", stage_ ) ) {
+ 		if ( Ftype1< double >::get( target, "stage", stage_ ) ) {
 			// Hacks all, to set up the recvFuncs in either direction.
 			procFunc_ = target->field( "processIn" )->recvFunc();
 			reinitFunc_ = target->field( "reinitIn" )->recvFunc();
